Move MeasureFragmentWidth from types to measurefragment
